requirements.txt
mysql-connector-python
faker
pandas
sqlalchemy
matplotlib
numpy
